refactored previous tests
